TENTATIVE CHECKIN  -- BLEEDING EDGE (only PF_ROX until better tested)
90KB of 'minor changes'... encompassing:

- a revised value indexing scheme that uses a more advanced, and better
  hash function (the earlier one did not handle text nodes well).

  the value index remains a single index that can power access to
  attribute nodes, text nodes, and (and this is special) element
  nodes (by their text value). This scheme is even amendable for
  updates, as it does not need to reconstruct full text values to
  update the index (ie hashvalues can be composed of the hash values
  of child nodes). Note: element indexing is not yet switched on.

  the hash function supports both by-string selections:

    doc('x')//a[@b = '42']

  as well as decimal/integer based access:

    doc('x')//a[@b = 42.00]

  it only supports equi-select (and join) and parsing/rounding problems
  make it practically meaningless to support this on doubles.

- to make the indexing work, it became apparent that double-based
  acces had to be dropped in favor of decimal-based access. However,
  this would only be meaningful if MonetDB/XQuery would support proper
  decimals.

  to enable this, pf_support.mx introduces a new MonetDB type "dec".
  It is a very simple decimal implementation 12d.6d  (i.e. a fixed
  decimal point at the 6th digit). The 18 digits fit a 64-bits integer.

  note: dec_values (mps variable) is no longer the same as dbl_values!!

- all value and element indices are again full indices. The business
  of pruning frequent values/elements has been dropped to allow easier
  experimentation in PF_ROX.

- adapted the heuristic.c Core-rewrites that introduce runtime choice-plans
  that use the velue indices. The changes had to be made due to the
  changes in the indexing scheme (above). Also, the heuristics are
  rewritten into a form that the algebra compiler can handle as well.

  what is missing still are the translations for the builtins
  (pf:text pf:attribute pf:idxfail pf:indexfailed)

- implemented doc-available()

+++ pathfinder/compiler/algebra/algebra.c
@@ -458,7 +458,7 @@ PFalg_lit_str (char *value)
 
 /** construct literal decimal (atom) */
 PFalg_atom_t
-PFalg_lit_dec (float value)
+PFalg_lit_dec (long long int value)
 {
     return (PFalg_atom_t) { .type = aat_dec, .val = { .dec_ = value } };
 }
